Convert all code to new Stream API.
Also made a number of improvmements to tt, making it compatible with more
shells.

+++ src/Io.c
@@ -134,7 +134,7 @@ int
 IoVprintf(Io * io, const char *fmt, va_list ap)
 {
     char *buf;
-    size_t write;
+    int write;
 
     int ret;
